Web client: New 'div' elements with IDs 'bar', 'barPosition', 'barCenter'
Elements 'bar', 'barPosition', 'barCenter' offer a view for the deviation from button center frequency.
It is the 'detune' value of the oscillator at each button. 'bar' represents the interval of possible values,
'barCenter' depicts the position at which detune is 0 and 'barPosition' corresponds to the current detune
value set by device sensors.

Changes
	- views/index.html, views/instrument.html:
		- 'div' elements added with IDs 'bar', 'barPosition' and 'barCenter'
	- public/stylesheet.css:
		- Rules for '#bar', '#barCenter' and '#barPosition' added.
	- public/javascript/functionsForPiece.mjs:
		- Function 'sensorBarListenerFunc' added.
		- Function 'sensorListenerFunc': the 'round' function is used to 'detune'.
		- Function 'initSensorsAndAttachListeners':
			- Argument 'document' added.
			- Returns an event listener function.
			- The 'sensorBarListenerFunc' is added to the 'reading' event listenerse.